Added a Policy to check if a user is an author or not of a post before deleting or editing it. Also, added another one to determine if a post is visible or not, and control if it can be seen or not.

+++ app/Http/Controllers/Admin/PostController.php
@@ -53,7 +53,7 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
-        // $this->authorize('author', $post);
+        $this->authorize('author', $post);
 
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
@@ -93,7 +93,7 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        // $this->authorize('author', $post);
+        $this->authorize('author', $post);
         $post->delete();
         // Cache::flush();
         return redirect()->route('admin.posts.index')->with('info', 'Post deleted succesfully!');
